+++ Assets/VRM10_Samples/VRM10Viewer/VRM10ViewerState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bbda0e60b1868ff4f95c09b15f35a35f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
